feat: added rules at services

+++ src/DevIO.Business/Services/SupplierService.cs
@@ -18,17 +18,54 @@ public class SupplierService : BaseService, ISupplierService
         if (!ExecuteValidation(new SupplierValidation(), supplier) ||
             !ExecuteValidation(new AddressValidation(), supplier.Address)) return;
 
+        if (_supplierRepository.SearchAsync(s => s.Document == supplier.Document).Result.Any())
+        {
+            Notifier("There is already a supplier with the informed document");
+            return;
+        }
+
         await _supplierRepository.AddAsync(supplier);
     }
 
     public async Task UpdateAsync(Supplier supplier)
     {
         if (!ExecuteValidation(new SupplierValidation(), supplier)) return;
+
+        if (_supplierRepository.SearchAsync(
+                s => s.Document == supplier.Document &&
+                s.Id != supplier.Id
+            ).Result.Any()
+        )
+        {
+            Notifier("There is already a supplier with the informed document");
+            return;
+        }
         await _supplierRepository.UpdateAsync(supplier);
     }
 
     public async Task RemoveAsync(Guid id)
     {
+        var supplier = await _supplierRepository.GetSupplierProductsAddressAsync(id);
+
+        if (supplier is null)
+        {
+            Notifier("Supplier does not exist");
+            return;
+        }
+
+        if (supplier.Products.Any())
+        {
+            Notifier("Supplier has registered products");
+            return;
+        }
+
+        var address = await _supplierRepository.GetAddressBySupplierAsync(id);
+
+        if (address is not null)
+        {
+            await _supplierRepository.RemoveAddressSupplier(address);
+        }
+
         await _supplierRepository.RemoveAsync(id);
     }
